Put recordDashboardClick back where it belongs
