+++ src/applications/personalization/dashboard/helpers.jsx
@@ -5,6 +5,7 @@ import { isPlainObject } from 'lodash';
 import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
 
 import { VA_FORM_IDS } from 'platform/forms/constants.js';
+import recordEvent from 'platform/monitoring/record-event';
 import getBenefitString from 'platform/forms-system/src/js/utilities/benefit-description';
 import getSavedFormDescription from 'platform/forms-system/src/js/utilities/form-description';
 
@@ -254,6 +255,17 @@ export function sipFormSorter(formA, formB) {
 
 export const isFormAuthorizable = formConfig =>
   !!formConfig && !!formConfig.authorize;
+
+export const recordDashboardClick = (
+  product,
+  actionType = 'view-link',
+) => () => {
+  recordEvent({
+    event: 'dashboard-navigation',
+    'dashboard-action': actionType,
+    'dashboard-product': product,
+  });
+};
 
 export const renderWidgetDowntimeNotification = (appName, sectionTitle) => (
   downtime,
